+++ react/RatingInline.tsx
@@ -11,8 +11,7 @@ import { withApollo } from 'react-apollo'
 import { ProductSummaryContext } from 'vtex.product-summary'
 import { ProductContext, Product } from 'vtex.product-context'
 import Stars from './components/Stars'
-import { generateBlockClass, BlockClass } from '@vtex/css-handles'
-import styles from './styles.css'
+import { useCssHandles } from 'vtex.css-handles'
 import TotalReviewsByProductId from '../graphql/totalReviewsByProductId.graphql'
 import AverageRatingByProductId from '../graphql/averageRatingByProductId.graphql'
 
@@ -63,10 +62,12 @@ const reducer = (state: State, action: ReducerActions) => {
   }
 }
 
-const RatingInline: FunctionComponent<BlockClass & Props> = props => {
-  const { blockClass, client } = props
+const CSS_HANDLES = ['inlineContainer'] as const
 
-  const baseClassNames = generateBlockClass(styles.inlineContainer, blockClass)
+const RatingInline: FunctionComponent<Props> = props => {
+  const { client } = props
+
+  const handles = useCssHandles(CSS_HANDLES)
   const { product }: ProductContext = useContext(ProductSummaryContext)
   const { productId }: Product = product || {}
 
@@ -109,7 +110,7 @@ const RatingInline: FunctionComponent<BlockClass & Props> = props => {
   }, [client, productId])
 
   return (
-    <div className={`${baseClassNames} review-summary mw8 center`}>
+    <div className={`${handles.inlineContainer} review-summary mw8 center`}>
       {!state.hasTotal || !state.hasAverage ? null : state.total == 0 ? null : (
         <Fragment>
           <span className="t-heading-5 v-mid">