+++ react/RatingSummary.tsx
@@ -10,8 +10,7 @@ import { NormalizedCacheObject } from 'apollo-cache-inmemory'
 import { withApollo } from 'react-apollo'
 import { ProductContext, Product } from 'vtex.product-context'
 import Stars from './components/Stars'
-import { generateBlockClass, BlockClass } from '@vtex/css-handles'
-import styles from './styles.css'
+import { useCssHandles } from 'vtex.css-handles'
 import TotalReviewsByProductId from '../graphql/totalReviewsByProductId.graphql'
 import AverageRatingByProductId from '../graphql/averageRatingByProductId.graphql'
 
@@ -62,10 +61,12 @@ const reducer = (state: State, action: ReducerActions) => {
   }
 }
 
-const RatingSummary: FunctionComponent<BlockClass & Props> = props => {
-  const { blockClass, client } = props
+const CSS_HANDLES = ['summaryContainer'] as const
 
-  const baseClassNames = generateBlockClass(styles.summaryContainer, blockClass)
+const RatingSummary: FunctionComponent<Props> = props => {
+  const { client } = props
+
+  const handles = useCssHandles(CSS_HANDLES)
   const { product }: ProductContext = useContext(ProductContext)
   const { productId }: Product = product || {}
 
@@ -108,7 +109,7 @@ const RatingSummary: FunctionComponent<BlockClass & Props> = props => {
   }, [client, productId])
 
   return (
-    <div className={`${baseClassNames} review-summary mw8 center`}>
+    <div className={`${handles.summaryContainer} review-summary mw8 center`}>
       {!state.hasTotal || !state.hasAverage ? (
         <Fragment>Loading reviews...</Fragment>
       ) : state.total == 0 ? null : (