+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- CSS Handles, including new `writeReviewContainer` handle
+
 ## [1.2.2] - 2020-01-30
 
 ### Fixed